2_Representation_Learning/requirements.txt
torch>=1.8.0
transformers>=4.10.0
datasets
faiss-cpu
pandas>=1.2.0
tqdm>=4.50.0
scikit-learn>=0.24.0
PyYAML>=5.0
sentence-transformers>=2.0.0
ranx>=0.3.0
numpy>=1.19.0